seperated accordian content
