+++ react-practice copy/trycontextapi/src/ReactTestDay12/Accordian.css
@@ -1,0 +1,56 @@
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+  }
+  
+  .accordion-icon {
+    border: solid currentcolor;
+    border-width: 0 2px 2px 0;
+    display: inline-block;
+    height: 8px;
+    pointer-events: none;
+    transform: translateY(-2px) rotate(45deg);
+    width: 8px;
+  }
+  
+  .accordion-icon-rotated {
+    transform: translateY(2px) rotate(-135deg);
+  }
+  
+  .rotated {
+    transform: translateY(-2px) rotate(-135deg);
+    border: solid currentcolor;
+    border-width: 0 2px 2px 0;
+    display: inline-block;
+    height: 8px;
+    pointer-events: none;
+    width: 8px;
+  }
+  
+  .accordian-style {
+    background-color: beige;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100vw;
+    gap: 2rem;
+    padding:1rem;
+  }
+  
+  .divcontent {
+    background-color: lightgrey;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 50vw;
+    gap: 0.3rem;
+    flex-direction: column;
+    border: 2px solid black;
+  }
+  
+  .Content {
+    text-align: center;
+  }
+  
+
